+++ src/controllers/auth/index.ts
@@ -1,4 +1,5 @@
 export * from './protect';
+export * from './forgotPassword';
 export * from './resetPassword';
 export * from './session';
 export * from './signin';
