+++ src/controllers/index.ts
@@ -1,4 +1,4 @@
 export * from './auth';
 export * from './errorController';
-export * from './auth/resetPassword';
+export * from './auth/forgotPassword';
 export * from './test';
