+++ src/routes/authRouter.ts
@@ -1,11 +1,11 @@
-import { resetPassword, resetPasswordComplete, signIn, signUp } from '@/controllers';
+import { forgotPassword, resetPassword, signIn, signUp } from '@/controllers';
 import { Router } from 'express';
 
 const router = Router();
 
 router.post('/signup', signUp);
-router.post('/password/reset/request', resetPassword);
-router.post('/password/reset', resetPasswordComplete);
+router.post('/password/forgot', forgotPassword);
+router.post('/password/reset', resetPassword);
 router.post('/signin', signIn);
 
 export { router as authRouter };
